requirement.txt
Django==2.1.7
django-revproxy==0.9.15
pytz==2018.9
urllib3==1.24.1
